Split API into submodules with tests
